Refactor instance description
